[fixed] Distance effects by precision loss at large distances

+++ common/src/main/java/de/mrjulsen/paw/event/ModCommonEvents.java
@@ -3,6 +3,7 @@ package de.mrjulsen.paw.event;
 import java.io.IOException;
 import java.util.Optional;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import de.mrjulsen.paw.config.ModCommonConfig;
@@ -46,7 +47,7 @@ public final class ModCommonEvents {
             for (WireGraph graph : WireGraphManager.getAll(level)) {
                 WiresApi.GENERIC_BLOCK.getAccessor(graph).ifPresent(x -> {
                     for (WireNode node : x.get(pos)) {
-                        graph.removeNode(node.getId(), new Vector3f(pos.getX() + 0.5f, pos.getY() + 0.5f, pos.getZ() + 0.5f), Optional.of(player));
+                        graph.removeNode(node.getId(), new Vector3d(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5), Optional.of(player));
                     }
                 });                
             }
