+++ common/src/main/java/de/mrjulsen/paw/util/collision/ICollisionProvider.java
@@ -2,12 +2,13 @@ package de.mrjulsen.paw.util.collision;
 
 import java.util.Optional;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 
 public interface ICollisionProvider {
-    Optional<RaycastHitResult> tryHit(Level level, BlockPos pos, Vector3f rayOrigin, Vector3f rayDirection);
+    Optional<RaycastHitResult> tryHit(Level level, BlockPos pos, Vector3d rayOrigin, Vector3d rayDirection);
 }
