+++ common/src/main/java/de/mrjulsen/wires/Wire.java
@@ -4,11 +4,12 @@ import java.util.Optional;
 
 import javax.annotation.Nullable;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import de.mrjulsen.wires.render.WireRenderData;
 
-public record Wire(String name, Vector3f pos, @Nullable WirePoints collisionData, @Nullable WireRenderData renderData) {
+public record Wire(String name, Vector3d pos, @Nullable WirePoints collisionData, @Nullable WireRenderData renderData) {
     public Optional<WirePoints> getCollisionData() {
         return Optional.ofNullable(collisionData());
     }
