+++ common/src/main/java/de/mrjulsen/wires/graph/data/node/GenericBlockNodeData.java
@@ -3,6 +3,7 @@ package de.mrjulsen.wires.graph.data.node;
 import java.util.Objects;
 import java.util.Optional;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import de.mrjulsen.mcdragonlib.DragonLib;
@@ -63,17 +64,17 @@ public class GenericBlockNodeData extends NodeData implements INodeDataBlock {
     
     @Override
     public WireNode getOrCreateNode(WireGraph graph) {
-        return graph.createNode(this, new Vector3f(pos.getX(), pos.getY(), pos.getZ()));
+        return graph.createNode(this, new Vector3d(pos.getX(), pos.getY(), pos.getZ()));
     }
     
     @Override
     public Optional<ConnectorDataProvider> getConnectorCustomData(WireGraph graph, CustomData customData, WireNode node, int pointIndex) {
-        return Optional.of(new BasicConnectorDataProvider(new Vector3f(attachPoint)));
+        return Optional.of(new BasicConnectorDataProvider(new Vector3d(attachPoint)));
     }
 
     @Override
-    public Vector3f toWorldPos(IWireGraph graph) {
-        return new Vector3f(getBlockPos().getX(), getBlockPos().getY(), getBlockPos().getZ()).add(getAttachPoint());
+    public Vector3d toWorldPos(IWireGraph graph) {
+        return new Vector3d(getBlockPos().getX(), getBlockPos().getY(), getBlockPos().getZ()).add(getAttachPoint());
     }
 
     @Override
